+++ ui/src/app/core/manage-plugins/manual-config/manual-config.component.ts
@@ -190,7 +190,7 @@ export class ManualConfigComponent implements OnInit {
   }
 
   public openFullConfigEditor() {
-    this.$router.navigate(['/config'])
+    void this.$router.navigate(['/config'])
     this.$activeModal.close()
   }