+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
   - the default view is the **new inline view**, which shows the differences together
   - a new button allows switching between the inline view and the **existing side-by-side view**
 - improvements to frontend route auth, guards and auth services
+- added config option to configure a webroot to access the UI from
 
 ### Other Changes